FIX: player can no longer jump on coin
changed the coin collider to a sensor. then, bc collisions no longer were registering, i did an intersection test instead (should be equivalent for this). the only thing i don't like is that the way i did it i had to use a `RapierContext`, which might not be good? not sure.